{"types":"C","runs":4}

+++ Problem17.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class Problem17 extends ConsoleProgram {
 	public void run() {
-		int d = readDouble("Enter number: ");
+		double d = readDouble("Enter number: ");
 		int integerPart = (int)d;
 		double fractionPart = d - integerPart;
 		println("integerPart: " + integerPart);
